Revert " Add Pagination to Contributions Table (#205)"
This reverts commit 8e9df561506feccf6153d32d1a634df19dff7ede.
