+++ apps/protocol-frontend/src/components/AttestationsTableShell.tsx
@@ -20,10 +20,9 @@ import MyAttestationsTable from './MyAttestationsTable';
 import ModalWrapper from './ModalWrapper';
 import { GovrnSpinner } from '@govrn/protocol-ui';
 import BulkAttestationModal from './BulkAttestationModal';
-import { useContributions } from '../contexts/ContributionContext';
 
 const AttestationsTableShell = () => {
-  const { isDaoContributionLoading, daoContributions } = useContributions();
+  const { isDaoContributionLoading, daoContributions } = useUser();
   const localOverlay = useOverlay();
   const { setModals } = useOverlay();
   const [selectedContributions, setSelectedContributions] = useState<any[]>([]);
