+++ apps/protocol-frontend/src/components/ProfileForm.tsx
@@ -9,11 +9,11 @@ import {
   linearFormValidation,
 } from '../utils/validations';
 import { ProfileFormValues } from '../types/forms';
-import { BASE_URL } from '../utils/constants';
+import { ValidationError } from 'yup';
 
 const LINEAR_CLIENT_ID = import.meta.env.VITE_LINEAR_CLIENT_ID;
 const LINEAR_REDIRECT_URI = import.meta.env.VITE_LINEAR_REDIRECT_URI;
-const BACKEND_ADDR = `${BASE_URL}`;
+const BACKEND_ADDR = `${import.meta.env.VITE_PROTOCOL_BASE_URL}`;
 
 const ProfileForm = () => {
   const { userData, updateProfile, disconnectLinear } = useUser();
