+++ apps/protocol-frontend/src/components/ReportForm.tsx
@@ -13,7 +13,6 @@ import { FormProvider, useForm, SubmitHandler } from 'react-hook-form';
 import { yupResolver } from '@hookform/resolvers/yup';
 import { reportFormValidation } from '../utils/validations';
 import { ContributionFormValues } from '../types/forms';
-import { useContributions } from '../contexts/ContributionContext';
 
 function CreateMoreSwitch({
   isChecked,
@@ -43,8 +42,13 @@ function CreateMoreSwitch({
 }
 
 const ReportForm = ({ onFinish }: { onFinish: () => void }) => {
-  const { allDaos, isUserActivityTypesLoading, userActivityTypes } = useUser();
-  const { isCreatingContribution, createContribution } = useContributions();
+  const {
+    isCreatingContribution,
+    isUserActivityTypesLoading,
+    userActivityTypes,
+    createContribution,
+    allDaos,
+  } = useUser();
 
   const localForm = useForm({
     mode: 'all',
