--- apps/protocol-frontend/src/components/TableEndMessage.tsx
@@ -1,11 +1,0 @@
-import { Text } from '@chakra-ui/react';
-
-const TableEndMessage = ({ message }: { message?: string }) => {
-  return (
-    <Text fontSize="sm" align={'center'} padding={3}>
-      {message ?? 'No more results'}
-    </Text>
-  );
-};
-
-export default TableEndMessage;
