--- apps/protocol-frontend/src/contexts/utils.ts
@@ -1,7 +1,0 @@
-/**
- * This type is used to remotely control a pagination from a component.
- */
-export type Pagination = {
-  next: () => void;
-  hasMore: boolean;
-};
