+++ apps/protocol-frontend/src/main.tsx
@@ -9,21 +9,18 @@ import { GovrnTheme } from '@govrn/protocol-ui';
 
 import Routes from './Routes';
 import { OverlayContextProvider } from './contexts/OverlayContext';
-import { ContributionsContextProvider } from './contexts/ContributionContext';
+import { UserContextProvider } from './contexts/UserContext';
 import { AuthContextProvider } from './contexts/AuthContext';
 import { wagmiClient, chains } from './utils/web3';
 
 import '@rainbow-me/rainbowkit/styles.css';
-import { UserContextProvider } from './contexts/UserContext';
 
 const App = () => (
   <WagmiConfig client={wagmiClient}>
     <RainbowKitProvider chains={chains}>
       <AuthContextProvider>
         <UserContextProvider>
-          <ContributionsContextProvider>
-            <Routes />
-          </ContributionsContextProvider>
+          <Routes />
         </UserContextProvider>
       </AuthContextProvider>
     </RainbowKitProvider>
