+++ apps/protocol-frontend/src/pages/ContributionDetail.tsx
@@ -19,7 +19,6 @@ import ContributionDetailShell from '../components/ContributionDetailShell';
 import NewUserView from '../components/NewUserView';
 import { GOVRN_MOTTO } from '../utils/constants';
 import { UIContribution } from '@govrn/ui-types';
-import { useContributions } from '../contexts/ContributionContext';
 
 const UserView = () => {
   return (
@@ -38,8 +37,7 @@ const UserView = () => {
 const ContributionDetails = () => {
   const { isConnected } = useAccount();
   const { id } = useParams();
-  const { userData } = useUser();
-  const { getContribution } = useContributions();
+  const { userData, getContribution } = useUser();
   const { isAuthenticated } = useAuth();
   const [contribution, setContribution] = useState<UIContribution>(
     {} as UIContribution,
