+++ package.json
@@ -78,7 +78,6 @@
     "react-dom": "17.0.2",
     "react-hook-form": "^7.27.1",
     "react-icons": "^4.3.1",
-    "react-infinite-scroll-component": "^6.1.0",
     "react-router-dom": "^6.2.2",
     "react-select": "^5.3.2",
     "react-table": "^7.7.0",
